Add drop game feature for users
Users can remove games from thier libraries

+++ database.py
@@ -576,12 +576,29 @@ class Database:
         if is_successful:
             game = self.get_game(game_id)
             statement = "INSERT INTO GAMES_OF_USERS(USER_ID, GAME_ID, TITLE, TIME_PURCHASED) VALUES(%s, %s, %s, CURRENT_DATE)"
-            data = (user_id, game_id, game.title,)
+            data = (user_id, game_id, game.title)
             query = statement, data
             self.query_database(query)
 
         self.disconnect()
         return is_successful
+
+    def remove_game_from_user(self, user_id, game_id):
+        self.connect()
+
+        statement = "SELECT * FROM GAMES_OF_USERS WHERE (GAME_ID = %s) AND (USER_ID = %s)"
+        data = (game_id, user_id,)
+        query = statement, data
+        self.query_database(query)
+
+        user_has_game = self.cursor.rowcount != 0
+        if user_has_game:
+            statement = "DELETE FROM GAMES_OF_USERS WHERE (USER_ID = %s) AND (GAME_ID = %s)"
+            data = (user_id, game_id)
+            query = statement, data
+            self.query_database(query)
+
+        self.disconnect()
 
     def update_game_favourite_variable(self, user_id, game_id, operation):
         self.connect()
@@ -1147,12 +1164,20 @@ class Database:
                         + " SET IS_FAVOURITE = FALSE" \
                         + " WHERE (USER1_ID = %s) AND (USER2_ID = %s)"
         elif operation == "REMOVE":
-            response = "Removed"
-            statement = "DELETE FROM FRIENDS WHERE (USER1_ID = %s) AND (USER2_ID = %s)"
+            are_friends = self.check_if_already_friends(user1_id, user2_id)
+            if are_friends:
+                response = "Removed"
+                statement = "DELETE FROM FRIENDS WHERE (USER1_ID = %s) AND (USER2_ID = %s)"
 
-            data = (user2_id, user1_id)
-            query = statement, data
-            self.query_database(query)
+                data = (user2_id, user1_id)
+                query = statement, data
+                self.query_database(query)
+
+                data = (user1_id, user2_id)
+                query = statement, data
+                self.query_database(query)
+
+            return response
 
         data = (user1_id, user2_id)
         query = statement, data
